# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API заявок, бригад, планирования и оперативного перепланирования.",
        "title": "Routing Planner API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/plans/{id}/events": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "replanning"
                ],
                "summary": "Применить событие и перестроить план",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID исходного плана",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Событие",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/replanning.EventDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/replanning.Result"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "plans.Plan": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "seed": {
                    "type": "integer"
                },
                "solution": {
                    "type": "object"
                },
                "solver_name": {
                    "type": "string"
                },
                "source_plan_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "running",
                        "feasible",
                        "partial",
                        "infeasible",
                        "error"
                    ]
                },
                "time_limit_seconds": {
                    "type": "number"
                }
            }
        },
        "plans.ReplanEvent": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payload": {
                    "type": "object"
                },
                "plan_id": {
                    "type": "string"
                },
                "processed_at": {
                    "type": "string"
                },
                "resulting_plan_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "replanning.EventDTO": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "object"
                },
                "type": {
                    "type": "string",
                    "example": "cancel_request"
                }
            }
        },
        "replanning.Result": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/plans.ReplanEvent"
                },
                "plan": {
                    "$ref": "#/definitions/plans.Plan"
                }
            }
        },
...
        "web.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation_error"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        }
    }
}